+++ src/components/common/sidebarComponent.js
@@ -39,7 +39,7 @@ const SidebarComponent = (props) => {
     }
     if (roleId && roleId === EMPLOYEE_ROLE) {
       const menu = appMenuItems?.filter(
-        (item) => [1, 2, 3, 4, 7, 9].includes(item.id)
+        (item) => [1, 2, 3, 4, 7, 8, 9].includes(item.id)
       );
       setAppMenuItemsData(menu);
     }
